python/requirements.txt
torch
onnxscript
zstandard
msgpack
boto3
python-dotenv